requirements.txt
boto3
psycopg2
python-dotenv
docker